Cleanup: Move some validation code
